Ereased chatRoutes because we are not communicating with the server to get the information from OpenAI. Move what was in chatRoutes to chat.js and connected that to chat.handlebars
