+++ routes/api/index.js
@@ -2,10 +2,10 @@ const express = require('express');
 const router = express.Router();
 const userRoutes = require('./user-routes');
 const reviewRoutes = require('./review-routes');
-const chatRoutes = require('./chatRoutes');
+// const chatRoutes = require('./chatRoutes');
 
 router.use('/users', userRoutes);
 router.use('/reviews', reviewRoutes);
-router.use('/chat', chatRoutes)
+// router.use('/chat', chatRoutes)
 
 module.exports = router;
