+++ routes/home-routes.js
@@ -11,12 +11,16 @@ router.get('/', (req,res)=> {
 
 router.get('/login', (req, res) => {
     // If the user is already logged in, redirect the request to another route
-    if (req.session.logged_in) {
+    if (req.session.loggedIn) {
       res.redirect('/profile');
       return;
     }
   
     res.render('login');
   });
+
+  router.get('/chat', (req, res) => {
+    res.render('chat');
+  })
 
 module.exports = router;
