--- views/homepage.handlebars
@@ -1,3 +1,0 @@
-
-
-<script src="/js/login.js"></script>
